CGM BT address shown in xdrip UI
The CGM Bluetooth address is a unique. xDrip4iOS writes the address to shared user defaults as soon as it connects to a CGM.
FAX reads this address and uses the device as heartbeat.

This commit makes the address visible in the UI and can be compared to the address visible in xDrip4iOS